+++ app/Http/Controllers/Api/v1/RouteSheetsApiController.php
@@ -3,10 +3,12 @@
 namespace App\Http\Controllers\Api\v1;
 
 use App\Domain\EntitiesServices\RouteSheetEntityService;
+use App\Domain\Enums\Abilities;
 use App\Http\Requests\Api\PaginatedSortedFilteredListRequest;
 use App\Http\Requests\Api\SaveRouteSheetRequest;
 use App\Models\RouteSheet;
 use Dingo\Api\Http\Response;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Validation\ValidationException;
 use Saritasa\Exceptions\InvalidEnumValueException;
 use Saritasa\Exceptions\NotImplementedException;
@@ -48,9 +50,12 @@ class RouteSheetsApiController extends BaseApiController
      *
      * @throws NotImplementedException
      * @throws InvalidEnumValueException
+     * @throws AuthorizationException
      */
     public function index(PaginatedSortedFilteredListRequest $request): Response
     {
+        $this->authorize(Abilities::GET, new RouteSheet());
+
         $filters = $request->getFilters([
             RouteSheet::COMPANY_ID,
             RouteSheet::ROUTE_ID,
@@ -94,7 +99,11 @@ class RouteSheetsApiController extends BaseApiController
      */
     public function store(SaveRouteSheetRequest $request): Response
     {
-        $routeSheet = $this->routeSheetService->store($request->getRouteSheetData());
+        $routeSheetData = $request->getRouteSheetData();
+
+        $this->authorize(Abilities::CREATE, new RouteSheet($routeSheetData->toArray()));
+
+        $routeSheet = $this->routeSheetService->store($routeSheetData);
 
         return $this->response->item($routeSheet, $this->transformer);
     }
@@ -113,6 +122,8 @@ class RouteSheetsApiController extends BaseApiController
      */
     public function update(SaveRouteSheetRequest $request, RouteSheet $routeSheet): Response
     {
+        $this->authorize(Abilities::UPDATE, $routeSheet);
+
         $this->routeSheetService->update($routeSheet, $request->getRouteSheetData());
 
         return $this->response->item($routeSheet, $this->transformer);
@@ -126,11 +137,12 @@ class RouteSheetsApiController extends BaseApiController
      * @return Response
      *
      * @throws RepositoryException
-     * @throws ValidationException
      * @throws Throwable
      */
     public function destroy(RouteSheet $routeSheet): Response
     {
+        $this->authorize(Abilities::DELETE, $routeSheet);
+
         $this->routeSheetService->destroy($routeSheet);
 
         return $this->response->noContent();
